漆ノ型
